Reorganize the files in SimCalorimetry/CastorTechTrigProducer and merge .h files
