--- SimCalorimetry/CastorTechTrigProducer/src/SealModule.cc
@@ -1,5 +1,0 @@
-#include "FWCore/Framework/interface/MakerMacros.h"
-#include "FWCore/PluginManager/interface/ModuleDef.h"
-#include "SimCalorimetry/CastorTechTrigProducer/src/CastorTTRecord.h"
-
-DEFINE_FWK_MODULE(CastorTTRecord);
